+++ lib/src/ui/form_step.dart
@@ -271,7 +271,7 @@ class RPUIFormStepState extends State<RPUIFormStep> {
             child: Text(
               locale?.translate(widget.formStep.questions[index].title) ??
                   widget.formStep.questions[index].title,
-              style: Theme.of(context).textTheme.titleLarge,
+              style: Theme.of(context).textTheme.titleMedium,
             ),
           ),
           Padding(
@@ -281,11 +281,7 @@ class RPUIFormStepState extends State<RPUIFormStep> {
               widget.formStep.questions[index].answerFormat,
             ),
           ),
-          Divider(
-              indent: 1,
-              endIndent: 1,
-              color: Theme.of(context).dividerColor,
-              thickness: 2)
+          const Divider(height: 24),
         ],
       ),
     );