+++ lib/src/ui/questions/integer_question_body.dart
@@ -78,10 +78,7 @@ class RPUIIntegerQuestionBodyState extends State<RPUIIntegerQuestionBody>
           errorText: _valid ? null : _errorMessage,
           focusedBorder: UnderlineInputBorder(
             borderSide: BorderSide(
-              color: (CupertinoTheme.of(context).primaryColor ==
-                      CupertinoColors.activeBlue)
-                  ? Theme.of(context).primaryColor
-                  : CupertinoTheme.of(context).primaryColor,
+              color: Theme.of(context).colorScheme.primary,
             ),
           ),
         ),